fix seg-fault in johnson_parallel_impl for large number of vertices

+++ graph/source/apsp/johnson.cpp
@@ -110,7 +110,7 @@ void johnson_parallel_impl(graph_t& gr, std::vector<double>& output) {
         std::vector<double> d(boost::num_vertices(G));
         boost::dijkstra_shortest_paths(G, s, boost::distance_map(d.data()));
         for (int v = 0; v < V; v++) {
-            output[s * V + v] = d[v] + h[v] - h[s];
+            output[static_cast<size_t>(s) * static_cast<size_t>(V) + static_cast<size_t>(v)] = d[v] + h[v] - h[s];
         }
     }
 }
